Split up JSON and content linting rules
